+++ src/main/scala/io/shiftleft/fuzzyc2cpg/passes/CMetaDataPass.scala
@@ -3,7 +3,6 @@ package io.shiftleft.fuzzyc2cpg.passes
 import io.shiftleft.codepropertygraph.Cpg
 import io.shiftleft.codepropertygraph.generated.{Languages, nodes}
 import io.shiftleft.fuzzyc2cpg.Defines
-import io.shiftleft.fuzzyc2cpg.Utils.getGlobalNamespaceBlockFullName
 import io.shiftleft.passes.{CpgPass, DiffGraph, KeyPool}
 
 /**
@@ -21,7 +20,7 @@ class CMetaDataPass(cpg: Cpg, keyPool: Option[KeyPool] = None) extends CpgPass(c
     def addAnyNamespaceBlock(diffGraph: DiffGraph.Builder): Unit = {
       val node = nodes.NewNamespaceBlock(
         name = Defines.globalNamespaceName,
-        fullName = getGlobalNamespaceBlockFullName(None)
+        fullName = CMetaDataPass.getGlobalNamespaceBlockFullName(None)
       )
       diffGraph.addNode(node)
     }
@@ -31,4 +30,17 @@ class CMetaDataPass(cpg: Cpg, keyPool: Option[KeyPool] = None) extends CpgPass(c
     addAnyNamespaceBlock(diffGraph)
     Iterator(diffGraph.build())
   }
+}
+
+object CMetaDataPass {
+
+  def getGlobalNamespaceBlockFullName(fileNameOption: Option[String]): String = {
+    fileNameOption match {
+      case Some(fileName) =>
+        s"$fileName:${Defines.globalNamespaceName}"
+      case None =>
+        Defines.globalNamespaceName
+    }
+  }
+
 }
